Create functionality in plotter to allow deletion of old plots
Previously the plotter would add additional plots below existing ones
rather than deleting the existing plot when new data is added or the
plot is updated. This forced the user of the api to access
the private plotter field and manually remove the plot and then to
create a new graph object each time the plot is changed. This update
makes the graph handle this deletion. It may also come in useful at
a later stage to have more control over plot changes from the
graph (e.g., to make use of d3 data joins).

Bug: chromium:866423
Change-Id: I5a3fa69b5dff4c24c05d8a941b9b43ab20815e8d