Dashboard - Push HistogramSet processing into the backend
Currently, this is done in the initial call by the client to
/add_histograms. AppEngine imposes a hard 60s limit to respond to FE
requests, and the uploaded datasets are only getting bigger, not
smaller.

Latest rendering.mobile uploads look like they'd roughly take
about 2 minutes of processing time now (guesstimate based on timings of
what's able to actually run).

This CL writes the dataset out to cloud storage and then queues a task
to do the actual validation and processing of the data. This gives us
more breathing room, since we get 10 mins to run on the backend.

The downside currently is that any validation and errors surfaced can
be communicated back to the client, but now they'll only appear in
logs. We could think about doing something like passing back an ID or
something from the initial /add_histograms call so that the client
could potentially look up the status of an upload later.

Bug: chromium:874856
Change-Id: I356b62a062b815bde98c4afbe9b2636cba62de56