Rename reportTemplateIds to reportTemplateInfos.
The variable `reportTemplateIds` caused some confusion for me when
working through adding Redux to reports. The variable holds an object of
information about a report template. The current name made me assume
that it was a unique identifying string or integer.

Renaming the variable to `reportTemplateInfos` helps ease this confusion
for me. Let me know what you think! :)

Bug: chromium:855252
Change-Id: Ia78a7df54f2017df43088e250f3f13a6df5cbb31