Add raised-button to V2SPA.
This dom-module is a simple drop-in replacement for paper-button. It offers no
functionality. It only contains CSS to draw a raised, rounded, blue div.
Callers are expected to listen for click and set tabindex if necessary.
https://material.io/design/components/buttons.html#theming

V2SPA is attempting to earn back user trust using many tools, but is not above
including calming blues and bean-bag-shaped buttons in that arsenal.

As a followup, chops-ui could be checked into third_party so V2SPA could use
chops-button instead if it would achieve the desired psychological effect.

Demo: https://v2spa-dot-chromeperf.appspot.com/

See also the CSS vars CL:
https://chromium-review.googlesource.com/c/catapult/+/1180346

Bug: catapult:#4461
Change-Id: Ia9bdd25aaf0d52c7b933bc61f19e7e6215f593ca