Restructure Long Term Health Tool Output File Structure.
In this CL, we've changed the output file structure of the long term health tool
so there will be a separate directory for each individual run of the long term
health tool, and inside which contains directories for each milestones. This
change will make our data analysis process easier. More details can be found in
https://docs.google.com/document/d/1lNoZhHdZ-Q-PcUzXSaP38yWM7ZimGiq3582rVQTSlNI/edit?usp=sharing

Some flag change is also made in this CL due to the change of `compatibility_mode` flag,
as in CL:1174721.

Bug: chromium:863390
Change-Id: I112122ee37c974b95b717dabb753ca4a399c735b