Update patch set 1
Patch Set 1:

Thanks Dave!

There are lots of things I love about the new plots:
- The histograms to the right, comparing values before/after, so useful!
- The numbers on each of the "jumps" showing the magnitude of each change.
- The red/blue highlighting, which are now also highlighted to match the detailed info on each of the two selected commits.

The main thing I do not like: The heatmap plot :( It's somehow less intuitive to read, for example this one:
https://dev-dtu-99b6e21e-dot-pinpoint-dot-chromeperf.appspot.com/job/16a1d234640000

Unless I've seen lots of other plots like this, it's hard to figure out exactly how to "read" it. My eyes are drawn to the thicker darker rectangles, which I can easily misrepresent as means (which I guess are the thinner lines?).

Also the heatmap seems, in a way, to be presenting almost the same information that the histogram to the right does, except "all at once" instead of individually for each pair of commits. I wonder whether it would be best to keep a line chart with error bars (as before) for the main chart, but still have access to the more detailed distribution when doing the pairwise comparisons.

Couple of other small comments (possibly part of a separate issue dealing with magnitudes):
- On the same graph linked above, would be great if units on the Y axis were in MiB (also the increments, instead of scientific notation).
- On https://dev-dtu-99b6e21e-dot-pinpoint-dot-chromeperf.appspot.com/job/14c27c0c640000 on the increments maybe we don't need that many decimal points?

Patch-set: 1
Reviewer: Gerrit User 1123010 <1123010@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>