Update patch set 1
Patch Set 1:

Wow, this looks so much cleaner!

It took me a few minutes to figure out how to read the side-by-side histograms on the side with interleaved bars. I was wondering what other visualizations were considered?
Some examples:
Results.html was planning to try subtracting bin counts like this: https://us.v-cdn.net/5022108/uploads/FileUpload/bb/a7aacf3c35c85ae572c90ea8719350.png
Combining colors might be easier to read: https://mpatacchiola.github.io/blog/2016/11/12/the-simplest-classifier-histogram-intersection.html
If you don't want to muddy the color code, there are other ways to overlap bars: https://peltiertech.com/overlapped-bar-chart-thinner-bars-in-front/

Also, a minor nit: when the metric is bigger-is-better, the heatmap bars for regressions are low and the difference size is floating way up at the top, so it looks like it's drifting without an anchor. It might be easier for me to connect the highlights with the data point if they were closer to the bars.

Patch-set: 1