Embed UMA graphs in metrics visualization tool
To make it easier for users to compare metrics between Telemetry and
UMA, embed relevant UMAs as iframes for each main step of the rendering
pipeline. Methods are provided to display both histograms and timeline
graphs. The OS of the histograms is used to find the correct UMA data,
and the channels default to dev and canary.

Bug: catapult:#4562
Change-Id: I8e96c434f909c1cd6919fa1b2b59fb4c2f4eb45f