Update patch set 2
Patch Set 2:

(1 comment)

This looks great, thanks for sharing with me! 

One high level comment to do with what you will: I know that this is based on the Material Design spec pretty directly, so I may not know what I'm talking about here, but I am a little concerned about the accessibility implications of using color as the main way to represent on/off state for a toggle button. Long-term, I think we might want to consider adding text labels on toggle buttons for on/off as well, so that the state is clearer to read, especially for colorblind users.

Patch-set: 2