Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> Looks great, though I do worry a bit about D3's "write-only language" tendencies.

Do you think those are inherent to the language, or because people commonly use it in that way? It would definitely be possible to implement these charts with a templating system, but I think Polymer 1.0 introduces too much boilerplate. Maybe it'll be better with lit-html?

It's also possible to do a hybrid approach: using d3 just to calculate the scales, but using lit-html to do the drawing.

Patch-set: 1
Reviewer: Gerrit User 1000093 <1000093@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>