Fix bug in key selectors.
This fixes a bug related to using d3 selectors. In particular,
sometimes keys which are used as class names contain
spaces which means they are registered as multiple
separate classes and so cannot later be retrieved with a
selector based on the original name.
This cl changes the encoding used to a simpler implementation to
avoid bugs related to selectors in the future.

Bug: chromium:866423
Change-Id: I5a8476cb4705b2c9916342495334273a9e519ec8