Parser for metrics & splitted components
Here is the new feature that allows the user to select a metric when you don't
know exactly which one but there are some areas of interest like the browser, the size etc.
These available ones are obtained like this:

* a new menu was added to make possible to choose by sub-components;
  a metric is made by several different components that are split like this:
  memory:browser:subprocess:probe:components:size.
  The user has to choose for each of them which are the ones he/she is more interested in.
  All the available options will be given and returned in the drop-down menu for further
  selection of stories and diagnostics.

Here the old single app component is splitted in its main components
to be easier to manage all of them.
- app_component (the one that manages all sub-components)
- menu_component (the menu that provides choosing metrics step by step)
- table_component (the table that display and allow to choose stories)

The metrics_parser.js splits the metric in its subcomponents: browser,
subprocess, probe, components and size. This way the user is able to choose
step by step the metric and will be provided with all the available
options that match.

The parser_test.js is a test for the parsing method. It receives some
metrics and returns some options for these ones depending on the
subcomponents.






Bug: chromium:866423
Change-Id: I1ddbc3485a419e8e5ae3fb457cf410efdd56408f