Dashboard - List out docs for alerts.
This prints a doc link for each suite in the list of alerts when a bug
is filed.

Will append a footer to the existing comment like:

system_health.memory_mobile - Documentation links:
  http://system_health_docs

loading.benchmark - Documentation links:
  http://foo/loading

etc


Additionally, have another cl up for Pinpoint to add a similar comment
when it finishes a job, to include the specific docs for the suite it
bisected.

Bug: chromium:865998
Change-Id: I7d5b1e20515323e21a40521832e4a6e9533dc480