Update patch set 4
Patch Set 4:

Hi mythri, I'm not sure if this is the best approach here. The advanatage is that it requires no changes to the external API (in particular the data input format). The disadvantage is that you will see the callback function is often called with undefined arguments which I worry is a bad pattern. Also, more importantly the way in which ids are kept track of (the index of data) may be quite brittle as it relies on data not being reordered (although I have another CL not submitted yet which does successfully link the dots to their traces so this does work). Let me know if you think these problems could be issues, thanks :)

Patch-set: 4
Reviewer: Gerrit User 1145189 <1145189@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1293423 <1293423@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>