Send analytics data with the Beacon API.
CURRENTLY BLOCKED BY chromium:438941

The Beacon interface is used to schedule an asynchronous and non-blocking
request to a web server. Beacon requests use the HTTP POST method and
requests typically do not require a response. Requests are guaranteed to
be initiated before a page is unloaded and they are run to completion,
without requiring a blocking request [1].

Beacon requests are prioritized to avoid competing with time-critical
operations and higher priority network requests.

Beacon requests are guaranteed to be initiated before page is unloaded
and are allowed to run to completion without requiring blocking requests
or other techniques that block processing of user interactive events.
[2]

There are two reasons to refactor google-analytics.js to use fetch:
 - Analytics should not compete with loading chart or report data.
 - The code is simplier

[1]: https://developer.mozilla.org/en-US/docs/Web/API/Beacon_API
[2]: https://w3c.github.io/beacon/#introduction

Bug: chromium:855252
Change-Id: Ib8cf091412e4ab774355013b211e840ec6bb7591