devil: refactor perf_control.py for upcoming N5X support
The plan is to add more actions for Nexus 5X in PerfControl. Adding
another conditional would reveal a bit too much duplication. Converting
to a declarative approach: a dictionary keyed by device.product_model
describing how to enable/disable benchmarking.

The test added covers only the N5. It is good illustrative purposes,
while the other devices are not common in the perf waterfall.

Bug: chromium:383566
Change-Id: I056c4ec906f1410470ab4a23ae36ba8a5a18931c