Flip axes for CFG.
This CL flips the axes on the cumulative frequency graph so that
memory used is displayed on the x axis and cumulative frequency
on the y axis. It also updates the color scheme to use the
array of colors provided by d3. This extends the number of
supported colors from 2 to 10.

Bug: chromium:866423
Change-Id: I7ba19b06764ac5b19884e06d26f19b72282cfb3d