[pinpoint_cli] Start pinpoint_cli script
Start with a very basic script, basically just integrating with the
services.Request forked off from soundwave. It makes a request to list
the jobs of the authenticated user, and dumps the json to stdout.

To keep things simple, remove from soundwave the possibility of
supplying a custom path for the user credentials (the default should
be fine for anyone using the script).

Bug: chromium:879526
Change-Id: Ia2b90cc3b4a94974b6d5b2d8413bd1379efac765