Update patch set 3
Patch Set 3:

Hi Mythri,
This CL allows a user to click on a bar for a story and display the line plot for that story. Ideally, this should also change the state of the table to reflect that the data displayed has changed (maybe deselect the other stories in the table), but I think Florina's changes may make large changes to the code related to the table so it'll make sense to implement this after that I think.

Patch-set: 3
Reviewer: Gerrit User 1145189 <1145189@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1293423 <1293423@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false