Update patch set 4
Patch Set 4:

> Patch Set 4: Code-Review+1
> 
> > Patch Set 3:
> > 
> > Hi Mythri,
> > This CL allows a user to click on a bar for a story and display the line plot for that story. Ideally, this should also change the state of the table to reflect that the data displayed has changed (maybe deselect the other stories in the table), but I think Florina's changes may make large changes to the code related to the table so it'll make sense to implement this after that I think.
> 
> Yes, I agree. May be we could add a TODO somewhere either near the code related to the table or near the code related to the graph so we can keep track. I also think, updating the table is less important than just being able to see the graph. One more question, once we go to the dot plot is it easy to have something to go back to the histogram plot? Either providing a link for back or someother way. I think, we can always get the plot by selecting from table again. So not high priority unless it is easy.

Yh i think it's useful to be able to get back to the bar charts. At the moment that kind of works for free as adding/removing a story on the table goes back to the bar chart, but it'll certainly make sense to have some kind of back button (it should be pretty easy I think). I think I'll wait for some of Florina's changes again (the change will depend on a lot of the state tracking variables and table code, some of which I think will be removed/restructured).

Patch-set: 4
Reviewer: Gerrit User 1293423 <1293423@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>