Update the UI: table structure
This CL updates the next parts in the UI: the averageSampleValue column
from table was removed; the pop up box that appeared after selecting
from tree-menu was removed; additional columns regarding the labels
were attached and now they are stable columns in the table; the id was
removed from the first column and now the checkboxes are merged with
the seond one;
now just of couple of diagnostics are stores (i.e. : traceUrls, labels,
stories, storySetRepeats and benchmarkStart in case a label does not
exist anymore).

Bug: chromium:866423
Change-Id: If40e50f5f151cf83d058f5639799c2bf5ca1ad34