Stacked bar plot integration
This CL integrates the stacked bar plot.

The source data for the plot is taken from the tree menu. After the
user selects the items from the menu, the table will display just
a couple of metrics that match. (first metric is the top level one)

The stacked bar plot works for multiple metrics, multiple stories
and a couple of labels.

For default displaying (just after applying the changes from tree menu)
it will show a stacked bar plot for: all the metrics that are given
after filtering upon tree-menu items, all the stories from the top
level metric from these ones and all the available labels.

In future, the user will be able to select/unselect other metrics,
stories or labels and to change the plot.


Bug: chromium:866423
Change-Id: I59541351fc29cdf7aa54bc4e9ea9b1efc4561644