Update patch set 6
Patch Set 6:

> Patch Set 5:
> 
> Just a naming suggestion, since these errors are conceptually different and come from different places, call them differently. For example:
> 
> blink.error.all
> blink.error.network
> v8.console.api
> v8.error.js
> 
> The fact that they all end up in the devtools console should not affect their naming.

This metric is an approximation of "is benchmark running ok?" and will be used for mostly monitoring/alerting. The main number is console:error:all. If it regresses, then it is a signal that our WPR archives are broken and have to be re-recorded. The distinction between blink/v8 or v8.console/v8.error would be too low-level. I am also not sure if we actually need  console:error:js and console:error:network (but I am curious what percentage of errors is coming from network).

Patch-set: 6
Reviewer: Gerrit User 1002405 <1002405@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>