Update html dimensions.
This CL makes some changes to the dimensions of the
page to improve usability. The overflow hidden style
is removed to allow horizontal scrolling when the table or
graph is wide (currently the excess is clipped). The height
style attribute of the menu is removed so that the page's
height is smaller, this makes it easier to use the vertical
scrollbar when looking through the table.

Bug: chromium:866423
Change-Id: I5531e67f57d852f08e452010c4bb00a2f159f394