Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Benoit---
> 
> Here are the catapult tracing extensions for java base.{odex,vdex} memory footprint metrics.
> 
> Note that I group the writable and readonly sections together; the writable ones probably aren't interesting for us (they seem to just be a page in all cases) but seems better to throw them in, in case they start exploding for some reason.
> 
> It's not quite the same as what native_lib_memory/java_code_pages_pss.py does---it's much more specific to odex and vdex---so LMK what you think about that.
> 
> Thx

Are we sure that the files are "base.{o,v}dex" on all versions of Android we care about?
This is my main concern, the rest looks fine (with the caveat that I don't know Javascript).

A writable page in the mapping would be... .data? This is quite surprising that it isn't much larger though, perhaps it's something else.

Patch-set: 1
Reviewer: Gerrit User 1132034 <1132034@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>