Orders stacking and fixes color assignment to ensure consistency.
This CL sorts the stacks by name before plotting so that they are
plotted in a consistent order. The purpose of this is to make
comparisons for the same metrics across different
bars easier. Also, it adds an explicit mapping of colors to stacks.
The previous approach does not work as it does not handle cases
where there is inconsistency in the stacks for a bar (e.g., if a
bar is missing a metric for some reason it would assign an incorrect
color).

Bug: chromium:866423
Change-Id: I7e28391ac1d17425c3eb9e4e7a3befc768d66280