Updates for the general features.
This CL covers this changes:

1) completely removed the top-level drop-down menu;
2) included a similar functionality inside the table (the hierarchy
will go like this: metrics, stories, storysetRepeats/labels. The old
hierarchy was like this: metrics, stories, typeOfDiagnostic,
anyDiagnostics, chosenPlot);
3) kept the change of plots as a global feature;
3) changed some UI details;

For a better usage of the tool, the dot plot and cumulative frequency
plot should be exploited from the main/default table, whereas the bar
char plot and the stacked plot should be used from the secondary table
(the one obtained after filtering the data with the tree menu).



Bug: chromium:866423
Change-Id: Icf54cce4199334804777d3390710dd6df520d2d2