Add functionality to generate the shard map for the given benchmark.
In this CL, I've extended the functionality of the `generate_perf_sharding`
script to generate the shard_map only for the given benchmark. The user needs
to provide the `benchmark_name` and `builder_name`, and the script will
generate the shard map. Users can also set the number of shards to use as well
as the output file path.

I've also fixed some minor style issues.

Bug: catapult:#
Change-Id: Ib42f204c4cacc11bceffc0147a80135d1722b03c