Add functions to include APK, generate isolate and upload it.
In this CL, we added function to move the APK to a path that will be included
in isolate. Also we added a function to generate isolate and a function to
upload them to the isolate server. These three functions will be used in later
CL to run the long term health job in swarming.

Bug: chromium:863390
Change-Id: I45ceb96007e0fb93074885f6623e6b98f375b92f