Added adb_timeout parameter to allow access command output on timeout
This is a follow-up after https://crrev.com/c/1217424. The usual timeout
parameter will generate a separate ReraiserThread that will simply terminate the
command without attempting to retrieve its output first and thus making it
impossible for the caller to print the command output on timeout to help
developers debug the issue.

R=jbudorick@chromium.org

Bug: chromium:872257
Change-Id: I1d754f227befb843d1cbe41abba5fa96dc175c08