Added adb_timeout parameter to allow access command output on timeout
This is a follow-up after https://crrev.com/c/1217424.

The usual timeout parameter will simply raise the reraiser_thread.TimeoutError
without attempting to retrieve the command output, whereas logic inside
adb.Shell actually tries to retrieve whatever output has been accumulated before
the timeout and adds it as a parameter to the exception making it impossible for
the caller to print the command output on timeout to help developers debug the
issue.

R=jbudorick@chromium.org

Bug: chromium:872257
Change-Id: I1d754f227befb843d1cbe41abba5fa96dc175c08