Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> > Patch Set 7:
> > 
> > > Patch Set 4:
> > > 
> > > (1 comment)
> > > 
> > > Ben, I'm interpreting your comment as Tibor's. Is that correct?
> > 
> > I don't think so.
> > histogramsByEventTitle's keys are trace event titles. If you change them, then it no longer supports traces generated by old versions of chrome such as telemetry's reference version. In order to support both old and new event titles, you could alias the Histogram objects in histogramsByEventTitle between two keys like this:
> > histogramsByEventTitle.set('VrShellGl::DrawFrame', histogramsByEventTitle.get('Vr.DrawFrame'));
> > Please also file a crbug to remove the aliases in a few months, and add a comment to this code linking to that crbug.
> 
> Would this be valuable for other branches of chromium? Given that the ToT trace events have already changed.

We generally try to keep metrics supporting old and new versions of chrome, not just ToT. Specifically, this enables the chromeperf dashboard to run benchmarks against ToT (the "test" build of chrome) as well as a reference build of chrome. This helps diagnose regressions -- if metrics for the reference build move at the same time as the test build, then it's probably due to a change in the bot such as OS version; but if metrics for the ref build stay the same while the test build regresses, then it's probably a real regression. Generally, there are various sub-teams in Chrome that occasionally do analysis across many versions in order to detect long-term trends in Chrome performance to find "death by a thousand cuts" scenarios.

Feel free to schedule a quick VC if that would help.

Patch-set: 7
Reviewer: Gerrit User 1115874 <1115874@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>