Update patch set 2
Patch Set 2:

oh I see, that would not be the intention.
Then i might make sense to change the destination of the "Analyze benchmark results" to point directly to the results2.html if the job is ready, or to the current results2 frame if not. WDYT?

Patch-set: 2
Reviewer: Gerrit User 1132654 <1132654@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>