Added parameter to keep output logs on test timeout
This is a follow-up after https://crrev.com/c/1217424.

This will add a flag that will start using a 5% timeout for the internal adb
command allowing the logs to be collected by a mechanism implemented in
adb_wrapper rather than being lost due to thread being preempted by the thread
watchdog. We implement this behind a flag to ensure that this does not break any
tests that are already close to the timeout, but eventually may choose to remove
the flag and apply this logic for all users of the API.

R=jbudorick@chromium.org

Bug: chromium:872257, chromium:891314
Change-Id: I1d754f227befb843d1cbe41abba5fa96dc175c08