Update patch set 15
Patch Set 15:

Had a quick look over it, and also followed some of the earlier discussion.

I'm not quite sure if this is the right approach. The main issue I have is that this seems to be increasing quite a bit the complexity of timeout/retry wrappers, options, etc. everywhere. I worry that it's going to make it harder to reason about (and the current state of things is already quite challenging!). It may also lead to unexpected or surprising behaviors due to interference between the different options.

I feel like this calls out for a design doc?

If I had to propose an alternative, would it be feasible to remove the timeout/retry wrapper from RunShellCommand (or adb.Shell?), and push the responsibility of that timeout all the way down to the cmd_helper module? Basically we would have two timeouts: A timeout for high level functions (e.g. almost everything in device_utils), methods which might involve executing several shell commands; and a second timeout to be enforced for each individual shell command?

Patch-set: 15
Reviewer: Gerrit User 1123010 <1123010@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>