Fixed charts usage
This CL fixes some particular and corner cases for charts usage (i.e.
to use by default bar plot when you end up with just one metric during
your analysis). These changes are aimed to create a more natural and
flexible flow of analysis without manually changing the parameters (i.e.
to specifically change the typeofPlot into barChart).

The data for guidValue is now parsed from the beginning without being
necessary to parse each time we need a specific item (it might be
arrays of String, arrays of Number, Number or String).

It also contains the integration for duration metrics hierarchy, but
this one was made just to see if the hierarchy is good.

Bug: chromium:866423
Change-Id: Iab8eb54205efe2b7f362c3f4d3856467fd732ae9