Add force commit to SharedPrefs
Adds the optional force_commit argument to SharedPrefs' Commit method,
which allows the caller to bypass SharedPrefs' automatic no-op if the
stored prefs are not dirty.

The use case for this is restoring modified preferences at a later time,
where the SharedPrefs' contents are not dirty, but the file on the
device has been changed by something else.

Bug: chromium:887579
Change-Id: I058a05e0746b46f8c6e68f1269ceecfaa21bd398