Add option-group for v2spa
This component displays a recursive tree of checkboxes for users to select
options. Branches in the tree use expand-buttons to allow hiding or showing
their children. Callers may group options manually, but most use
OptionGroup.groupValues() to normalize sets of options like ['a:b','a:c'] into
trees like {a:['b','c']}.
If 'a' is an option along with 'a:b', then the 'a' checkbox exhibits tri-state
logic to allow selecting either just 'a' or 'a' plus all its descendents.

Demo: most dropdown-inputs on https://v2spa-dot-chromeperf.appspot.com/

Bug: catapult:#4461
Change-Id: I72daf6e03faba0bacafab83df98c6df8b6d542ac