trace-viewer: Add a close-button to hide a process.
Add a 'x' button on the process-header to hide that process. The process
can be shown again from the 'Processes' filter-button at the top-bar.

NOTE: the checkbox in the 'Processes' dropdown does not update if the
process is hidden using the new 'x' button. How should that be fixed?
Should the ProcessTrack emit some signal or something?

BUG=catapult:#4553

Change-Id: Ib271eb99299f9cc8dd587c33d25e4ed2495f36bf