Tree menu for memory and duration benchmarks
This CL integrates both memory and duration benchmarks in a single
tree menu. Depending on the data type, the user is able to choose
between multiple options.
Therefore the flow of the tree-menu will be one for all types.
(the previous version had one menu for each type)

I see potential modifications for more flexibility of this menu.


Bug: chromium:866423
Change-Id: I8fadd3aa46ae06bdd56a3769994b9e67c6d7ecd8