Update patch set 15
Patch Set 15:

(1 comment)

> Patch Set 14:
> 
> (1 comment)
> 
> I still see gerrit complaining about the tab spacing.

Well, I ran gofmt on the files to fix the formatting. Is the ask to completely replace tabs with 4 spaces? I ask because the rest of the file, as well as all the other golang files in the project use tabs instead of space.

Patch-set: 15
Reviewer: Gerrit User 1251641 <1251641@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>