trace-viewer: Fix crash in timeline process filter.
Fix an unhandled exception that occurs in the process filter menu
when opening a trace with kernel thread scheduling info, such as
"tracing/test_data/android_systrace.html"

These traces have a "Kernel" process track with does not have a
"process" member, but still has a "processBase" member.

Use the views' "processBase" member instead of "process" member
in order to avoid crashing when creating the check-boxes for
process filtering.

Change-Id: I294809970879e4d97be30b302461476a2270131f