timeline_view: Allow combined systraces
When we combine Chrome trace with Android systrace, one view is like:
  div class="track process-track-base"

This gets selectes as processViews(), but for some magical reasons
view.process is undefined. Seeing this nednguyen@ suggested this
workaround. I am not sure why it works and how to re-create this
situation in a unittest. Suggestions are welcome.

Bug: chromium:889048
Change-Id: I765abee12d11e11b8b085030542e4b813aadef70