Telemetry: show metric averages only in perf dash
To be similar to legacy metrics, we make TBMv2 metrics report only
average values to the perf dashboard so that when we switch to the
new code, we don't add unnecessary _count, _min, .... versions to
benchmarks that use these metrics. An exception is the frame_lengths
metric for which we also report the maximum value. So,
frame_lengths_max will replace existing max_frame_delay.

Bug: chromium:890757
Change-Id: Id55e80a848a651d57c1ee8fee0aec788678b851c