Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> > Patch Set 6:
> > 
> > LGTM, thanks.
> > 
> > I think you're right about the 100x change.  It means the continuity will be a bit weird after this change.  The _tbmv2 metrics have been in since August, so it may make sense to keep them named _tbmv2 for now, and later migrate the data to remove the suffix?
> 
> Thanks!
> 
> We have 1-month history of _tbmv2 metrics in rendering benchmarks, but, _tbmv2 metrics are filtered out in other smoothness benchmarks (e.g. smoothness.gpu_rasterization.tough_pinch_zoom_cases):
> 
> https://cs.chromium.org/chromium/src/tools/perf/measurements/smoothness.py?l=82
> 
> Ben, is it hard to divide historic data in the dashboard by a constant, before a certain date? And, generally, what do you recommend in this situation?

Would it be acceptable to delete the incorrect data and move forward without it?
If not, it would probably be possible to modify the data in the data store using dev_console. If you think that might be necessary, I can check with Simon.

Patch-set: 6
Reviewer: Gerrit User 1115874 <1115874@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>