scroll: Allow for more errors.
Changing the frequency of scroll-event dispatch mean there are more
events being generated, and thus being queued/coalesced. So it is
possible for the scroll to not have reached the goal by the time the
test terminates. So allow for a little more distance from the target
for the slow-wheel unit-test.

BUG=892036

Change-Id: Ifc7a6f79c51a15800fcd702d076237a4ef764049