Telemetry: migrate frame_times & percentage_smooth
This migrates frame_times_tbmv2 to frame_times and
smooth_frames to percentage_smooth. The rename for
smooth_frames is for continuity of graphs in perf dashboard.
We can rename the metric if we want later and migrate data.

This migration also fixes crbug.com/892385

To compare old/new values:
https://v2spa-dot-chromeperf.appspot.com/#session=2ba522d010d5e4a81b1bfa24be6d8d5a2e29ff6d49af15721377314c49a4ea1d

Bug: chromium:890757
Bug: chromium:892385
Change-Id: If5411c66b753ccaaa62543e886c423f338667619