Update patch set 2
Patch Set 2:

(CCing vmiura for opinion)

1- This affects frame_times and similar metrics, too. Shouldn't we ignore frames outside the animation segment for frame times? For example, in the pinpoint job you ran for this CL, it seems that we have considerably more long (>= 50ms) frames with the CL than without the CL (1117 vs 361). Look at frame_times_tbmv2 not frame_times, because frame_times on mobile has a bug (crbug.com/892385) right now.

2- Are you sure that the events you care about will always be in the interaction record or they may come after that too?

It might be better to keep segments (in rendering metrics) to show the part of the trace that we expect to be animation, as precisely as we can, and instead look for later events in metric-specific code, if they need to (i.e. expand the search for events in pipeline.html:91). WDYT?

Patch-set: 2
CC: Gerrit User 1116061 <1116061@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1112474 <1112474@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>