Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (CCing vmiura for opinion)
> 
> 1- This affects frame_times and similar metrics, too. Shouldn't we ignore frames outside the animation segment for frame times? For example, in the pinpoint job you ran for this CL, it seems that we have considerably more long (>= 50ms) frames with the CL than without the CL (1117 vs 361). Look at frame_times_tbmv2 not frame_times, because frame_times on mobile has a bug (crbug.com/892385) right now.
> 
> 2- Are you sure that the events you care about will always be in the interaction record or they may come after that too?
> 
> It might be better to keep segments (in rendering metrics) to show the part of the trace that we expect to be animation, as precisely as we can, and instead look for later events in metric-specific code, if they need to (i.e. expand the search for events in pipeline.html:91). WDYT?

Hmm, yeah I think that code for limiting the considered bounds to just the gesture event bounds was added to reduce noise.  Increasing it to the interaction bounds will give some erroneous bad frames.  Also as @chiniforooshan mentions, it will still not guarantee that we capture all of the segments related to the gesture.

Could we tag frames in the pipeline, so that we can more accurately filter the needed events?   (FrameTiming was meant to replace this)

Patch-set: 2